Move the trace loggers to a Loggers folder in EventLogExpert.Logging
