+++ src/EventLogExpert.Logging/Routing/LogSourceFactory.cs
@@ -2,6 +2,7 @@
 // // Licensed under the MIT License.
 
 using EventLogExpert.Logging.Abstractions;
+using EventLogExpert.Logging.Loggers;
 
 namespace EventLogExpert.Logging.Routing;
