+++ tests/Unit/EventLogExpert.Logging.Tests/Loggers/DispatchingTraceLoggerTests.cs
@@ -2,10 +2,10 @@
 // // Licensed under the MIT License.
 
 using EventLogExpert.Logging.Abstractions;
-using EventLogExpert.Logging.Routing;
+using EventLogExpert.Logging.Loggers;
 using Microsoft.Extensions.Logging;
 
-namespace EventLogExpert.Logging.Tests.Routing;
+namespace EventLogExpert.Logging.Tests.Loggers;
 
 public sealed class DispatchingTraceLoggerTests
 {
